refactor(blender): drop five more ignores after testing each against Blender
After auditing every rule left in ``blender_mmgpy/**/*.py``, only two
are genuinely Blender-API-mandated. The other five were avoidable:

- ``RUF012``: declare ``bl_options`` with a ``ClassVar[set[str]]``
  annotation. Runtime value is still a ``set`` (Blender's own
  requirement), so ``register_class()`` is unaffected.
- ``ARG002``: rename the unused ``event`` / ``context`` parameters
  on ``invoke()`` / ``draw()`` to ``_event`` / ``_context``. Blender
  calls these callbacks positionally, so the rename does not break
  the operator / preferences contract.
- ``RUF067``: move ``bl_info`` plus the registration helpers out of
  ``__init__.py`` into a new ``_register.py`` and re-export them.
  Blender resolves ``register`` / ``unregister`` / ``bl_info`` as
  attributes of the package, so the re-export is transparent.
- ``N802``: never actually fired — all our method names (``poll``,
  ``execute``, ``invoke``, ``draw``) are already lowercase.

What remains, with the reason each cannot be addressed in code:

- ``N801``: ``bpy.utils.register_class()`` rejects operators / panels
  unless the class name matches ``UPPER_TYPE_lowercase``
  (``MMGPY_OT_remesh``, ``MMGPY_PT_main_panel``); the registration
  helper validates the name against ``cls.bl_rna.identifier`` and
  raises on a mismatch.
- ``TC002``: ``bpy.props.<Property>`` callables only appear in class
  annotations (``radius: FloatProperty(default=0.1)``) but Blender's
  add-on metaclass evaluates those annotations at registration to
  materialise RNA property descriptors — so the names must stay in
  module globals. ``flake8-type-checking.exempt-modules`` would be a
  surgical fix but it is a global setting; tweaking it disrupts
  preview TC rules already passing in ``src/mmgpy/``, so the
  per-file ignore is the smaller blast radius.
